packages/mixinv2-examples/pyproject.toml
[project]
name = "mixinv2-examples"
version = "0.0.0.dev0"
description = "Example case studies for MIXINv2"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
  "mixinv2",
  "mixinv2-library",
  "starlette",
  "uvicorn",
  "aiosqlite",
]

[project.scripts]
mixinv2-example = "mixinv2_examples._cli:main"

[dependency-groups]
dev = [
  "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
sources = { mixinv2 = { workspace = true }, mixinv2-library = { workspace = true } }
